+++ src/flavors/pboakland/static/js/views/legend-view.js
@@ -1,0 +1,15 @@
+var LegendView = require('../../../../../base/static/js/views/legend-view.js');
+
+module.exports = LegendView.extend({
+  render: function() {
+    var self = this,
+        data = _.extend({
+          items: this.options.config.items
+        }, Shareabouts.stickyFieldValues);
+
+    // use prepend here so as to not overwrite activity stream
+    this.$el.prepend(Handlebars.templates['legend'](data));
+
+    return this;
+  }
+});
