+++ src/flavors/pboakland/templates/index.html
@@ -57,8 +57,4 @@
   Analytics, custom JS, and such go here
  -->
 {% block includes %}
-<script src="{{STATIC_URL}}js/views/pboakland-app-view.js"></script>
-<script src="{{STATIC_URL}}js/views/pboakland-place-detail-view.js"></script>
-<script src="{{STATIC_URL}}js/views/pboakland-sidebar-view.js"></script>
-<script src="{{STATIC_URL}}js/views/pboakland-legend-view.js"></script>
 {% endblock %}
